feat: show events; show week dates in mobile view
Given an id, title, start time and duration, the calendar events will be
displayed on the calendar page. As of now, the calendar events are
mocked to be scheduled on the current day of the week. In future, this
can easily be modified. Additionally, the mobile view now reflects the
current week.