feat: implement calendar week navigation
This is a very rudimentary approach taken. It can work for most cases
but edge cases have not been considered. For example, the day of the
month may sometimes be correct and other times incorrect. We can
implement that a little later down the line. In this commit, we also
minify the three-dot dropdown menu using shadcn/ui components and
removing our dependence on Tailwind's Headless UI and Heroicons.