chore: update brand icon; add new sections
The brand icon was updated to the first calendar icon found online. In
future, if things take off we can change this to something that's
better. Additionally, we now have new sections in README.md: "Repository
Activity", "Star History", and login credentials for the demo. There is
also a new table of contents and motive for the app.